+++ src/apyarray.h
@@ -202,11 +202,8 @@ public:
 
             ARRAY_TYPE result
                 = static_cast<const ARRAY_TYPE*>(this)->create_array(new_shape);
-            std::copy_n(
-                std::begin(_data) + idx * _itemsize * element_stride,
-                _itemsize * element_stride,
-                std::begin(result._data)
-            );
+            auto src = std::begin(_data) + idx * _itemsize * element_stride;
+            std::copy_n(src, _itemsize * element_stride, std::begin(result._data));
             return RESULT_TYPE(result);
         }
     }
@@ -231,11 +228,8 @@ public:
             if (std::holds_alternative<nb::int_>(element)) {
                 auto idx = static_cast<std::ptrdiff_t>(std::get<nb::int_>(element));
                 idx = adjust_integer_index(idx, dim, "__getitem__");
-                std::copy_n(
-                    input_it + _itemsize * idx * strides[dim],
-                    _itemsize * strides[dim],
-                    output_it
-                );
+                auto src = input_it + _itemsize * idx * strides[dim];
+                std::copy_n(src, _itemsize * strides[dim], output_it);
                 return strides[dim];
             } else { /* std::holds_alternative<nb::slice>(element) */
                 auto&& slice = std::get<nb::slice>(element);
@@ -243,20 +237,16 @@ public:
                 std::size_t elements_copied = 0;
                 if (step < 0) {
                     for (std::ptrdiff_t src_i = start; src_i > stop; src_i += step) {
-                        std::copy_n(
-                            input_it + _itemsize * src_i * strides[dim],
-                            _itemsize * strides[dim],
-                            output_it + _itemsize * elements_copied
-                        );
+                        auto src = input_it + _itemsize * src_i * strides[dim];
+                        auto dst = output_it + _itemsize * elements_copied;
+                        std::copy_n(src, _itemsize * strides[dim], dst);
                         elements_copied += strides[dim];
                     }
                 } else { /* step >= 0 */
                     for (std::ptrdiff_t src_i = start; src_i < stop; src_i += step) {
-                        std::copy_n(
-                            input_it + _itemsize * src_i * strides[dim],
-                            _itemsize * strides[dim],
-                            output_it + _itemsize * elements_copied
-                        );
+                        auto src = input_it + _itemsize * src_i * strides[dim];
+                        auto dst = output_it + _itemsize * elements_copied;
+                        std::copy_n(src, _itemsize * strides[dim], dst);
                         elements_copied += strides[dim];
                     }
                 }
@@ -395,9 +385,10 @@ public:
             }
         }
 
-        // Extract resulting ndim and chunk size
+        // Extract resulting ndim and chunk size. The chunk size is the number of
+        // contiguous elements to copy for each `true` value in the slice array.
         std::size_t res_ndim = _ndim - key.ndim() + 1;
-        std::size_t chnk
+        std::size_t chunk_size
             = fold_shape(std::end(_shape) - res_ndim + 1, std::end(_shape));
 
         // Compute the resulting shape
@@ -405,18 +396,14 @@ public:
         std::copy_n(std::end(_shape) - res_ndim, res_ndim, std::begin(res_shape));
         res_shape[0] = std::count(key.data(), key.data() + key.size(), true);
 
-        // Create the resulting flattened array
-        ARRAY_TYPE res = static_cast<const ARRAY_TYPE*>(this)->create_array(res_shape);
-
         // Copy elements into result
+        ARRAY_TYPE res = static_cast<const ARRAY_TYPE*>(this)->create_array(res_shape);
         std::size_t i_dst = 0;
         for (std::size_t i = 0; i < key.size(); i++) {
             if (key.data()[i]) {
-                std::copy_n(
-                    std::begin(_data) + i * chnk * _itemsize,
-                    chnk * _itemsize,
-                    std::begin(res._data) + i_dst++ * chnk * _itemsize
-                );
+                auto src = std::begin(_data) + i * chunk_size * _itemsize;
+                auto dst = std::begin(res._data) + i_dst++ * chunk_size * _itemsize;
+                std::copy_n(src, chunk_size * _itemsize, dst);
             }
         }
 
@@ -489,11 +476,8 @@ public:
             if (std::holds_alternative<nb::int_>(element)) {
                 auto idx = static_cast<std::ptrdiff_t>(std::get<nb::int_>(element));
                 idx = adjust_integer_index(idx, dim, "__setitem__");
-                std::copy_n(
-                    input_it,
-                    _itemsize * strides[dim],
-                    output_it + _itemsize * strides[dim] * idx
-                );
+                auto dst = output_it + _itemsize * strides[dim] * idx;
+                std::copy_n(input_it, _itemsize * strides[dim], dst);
                 return strides[dim];
             } else { /* std::holds_alternative<nb::slice>(element) */
                 auto&& slice = std::get<nb::slice>(element);
@@ -501,20 +485,16 @@ public:
                 std::size_t elements_copied = 0;
                 if (step < 0) {
                     for (std::ptrdiff_t dst_i = start; dst_i > stop; dst_i += step) {
-                        std::copy_n(
-                            input_it + _itemsize * elements_copied,
-                            _itemsize * strides[dim],
-                            output_it + _itemsize * dst_i * strides[dim]
-                        );
+                        auto src = input_it + _itemsize * elements_copied;
+                        auto dst = output_it + _itemsize * dst_i * strides[dim];
+                        std::copy_n(src, _itemsize * strides[dim], dst);
                         elements_copied += strides[dim];
                     }
                 } else { /* step >= 0 */
                     for (std::ptrdiff_t dst_i = start; dst_i < stop; dst_i += step) {
-                        std::copy_n(
-                            input_it + _itemsize * elements_copied,
-                            _itemsize * strides[dim],
-                            output_it + _itemsize * dst_i * strides[dim]
-                        );
+                        auto src = input_it + _itemsize * elements_copied;
+                        auto dst = output_it + _itemsize * dst_i * strides[dim];
+                        std::copy_n(src, _itemsize * strides[dim], dst);
                         elements_copied += strides[dim];
                     }
                 }
@@ -580,6 +560,7 @@ public:
             );
             throw nb::value_error(error_msg.c_str());
         }
+        assert(_itemsize = val._itemsize);
 
         // Compute the slice shape
         std::vector<std::size_t> slice_shape = compute_slice_shape(key);
@@ -646,12 +627,204 @@ public:
         set_item_from_array(key, array_val);
     }
 
+    //! Set items in `*this` indexed by a boolean ndarray. This function assumes that
+    //! both `key` is of appropriate shapes to perform the assignment
+    void set_item_ndarray_from_scalar(
+        const nb::ndarray<bool, nb::c_contig>& key,
+        const scalar_variant_t<ARRAY_TYPE>& val
+    )
+    {
+        // Make sure that all bit specifiers in `*this` and `val` are equal.
+        if (!static_cast<const ARRAY_TYPE*>(this)->is_same_spec(val)) {
+            std::string error_msg = fmt::format(
+                "{}.__setitem__: `val` has different bit specifiers than `self`",
+                ARRAY_TYPE::ARRAY_NAME
+            );
+            throw nb::value_error(error_msg.c_str());
+        }
+
+        // Extract resulting ndim and chunk size. The chunk size is the number of
+        // contiguous elements to copy for each `true` value in the slice array.
+        std::size_t res_ndim = _ndim - key.ndim() + 1;
+        std::size_t chunk_size
+            = fold_shape(std::end(_shape) - res_ndim + 1, std::end(_shape));
+
+        // Set values in `*this` array
+        for (std::size_t i = 0; i < key.size(); i++) {
+            if (key.data()[i]) {
+                for (std::size_t j = 0; j < chunk_size; j++) {
+                    std::size_t offset = i * chunk_size * _itemsize + j * _itemsize;
+                    val.copy_n_to(std::begin(_data) + offset, _itemsize);
+                }
+            }
+        }
+    }
+
+    //! Set items in `*this` indexed by a boolean ndarray. This function assumes that
+    //! both `key` and `val` are of appropriate shapes to perform the assignment
+    void set_item_ndarray_from_array(
+        const nb::ndarray<bool, nb::c_contig>& key, const ARRAY_TYPE& val
+    )
+    {
+        // Make sure that all bit specifiers in `*this` and `val` are equal.
+        if (!static_cast<const ARRAY_TYPE*>(this)->is_same_spec(val)) {
+            std::string err_msg = fmt::format(
+                "{}.__setitem__: `val` has different bit specifiers than `self`",
+                ARRAY_TYPE::ARRAY_NAME
+            );
+            throw nb::value_error(err_msg.c_str());
+        }
+        assert(_itemsize == val._itemsize);
+        assert(_shape.size() > 0);
+        assert(val._shape.size() > 0);
+
+        // Compute the slice `ndim`
+        std::size_t slice_ndim = _ndim - key.ndim() + 1;
+        if (slice_ndim == 1 && val._ndim > 1) {
+            // Special case for Numpy slicing compatibility. Only when the slice ndim is
+            // one and `val` ndim is greater than one, do we throw a `TypeError` here.
+            // If the slice ndim is greater than one, we might first need to squeeze of
+            // dimensions of `val` before determening success or failure.
+            std::string err_msg = fmt::format(
+                "{}.__setitem__: ndarray boolean slicing requires 1-dimensional input, "
+                "but input has {} dimensions",
+                ARRAY_TYPE::ARRAY_NAME,
+                val._ndim
+            );
+            throw nb::type_error(err_msg.c_str());
+        }
+
+        // Count of number of `true` values of the key array
+        std::size_t n_true = std::count(key.data(), key.data() + key.size(), true);
+
+        // Compute the slice shape
+        std::vector<std::size_t> slice_shape(slice_ndim);
+        slice_shape[0] = std::max(n_true, std::size_t(1));
+        std::size_t ndim_cpy = slice_ndim - 1;
+        std::copy_n(std::end(_shape) - ndim_cpy, ndim_cpy, std::begin(slice_shape) + 1);
+
+        // The chunk size, i.e., the number of elements to copy for each `true` value in
+        // the slice array
+        std::size_t n_chunk = fold_shape(std::end(_shape) - ndim_cpy, std::end(_shape));
+
+        // Start and sentinel iterator pair of `*this` shape squeezed off leading ones
+        auto squeeze_end = std::end(val._shape);
+        auto squeeze_begin = std::find_if(
+            std::begin(val._shape), std::end(val._shape) - 1, [](auto dim) {
+                return dim != 1;
+            }
+        );
+
+        bool is_squeezed_shape_equal = std::equal(
+            std::begin(slice_shape), std::end(slice_shape), squeeze_begin, squeeze_end
+        );
+        if (val._shape == slice_shape || is_squeezed_shape_equal) {
+            // The slice exactly matches the shape of `val`. Simply copy the data
+            std::size_t slice_i = 0;
+            for (std::size_t i = 0; i < key.size(); i++) {
+                if (key.data()[i]) {
+                    auto src = std::begin(val._data) + slice_i++ * n_chunk * _itemsize;
+                    auto dst = std::begin(_data) + i * n_chunk * _itemsize;
+                    std::copy_n(src, n_chunk * _itemsize, dst);
+                }
+            }
+            return;
+        }
+
+        // Create vector of the squeezed `val` shape
+        std::vector<std::size_t> squeeze_shape(squeeze_begin, squeeze_end);
+
+        if (is_broadcastable(squeeze_shape, slice_shape)) {
+            // The elements of `val` can be broadcast to match the slice
+            vector_type broadcast(_itemsize * n_chunk * slice_shape[0]);
+            broadcast_data_copy(
+                std::begin(val._data), // src
+                std::begin(broadcast), // dst
+                squeeze_shape,         // src_shape
+                slice_shape,           // dst_shape
+                _itemsize              // itemsize
+            );
+            std::size_t slice_i = 0;
+            for (std::size_t i = 0; i < key.size(); i++) {
+                if (key.data()[i]) {
+                    auto src = std::begin(broadcast) + slice_i++ * n_chunk * _itemsize;
+                    auto dst = std::begin(_data) + i * n_chunk * _itemsize;
+                    std::copy_n(src, n_chunk * _itemsize, dst);
+                }
+            }
+            return;
+        }
+
+        // The shape of `val` is not compatible with the slice shape
+        std::string err_msg = fmt::format(
+            "{}.__setitem__: shape mismatch; value array of shape {} can not be "
+            "broadcast to the required index slice shape {}",
+            ARRAY_TYPE::ARRAY_NAME,
+            tuple_string_from_vec(val._shape),
+            tuple_string_from_vec(slice_shape)
+        );
+        throw nb::value_error(err_msg.c_str());
+    }
+
+    //! Set items in `*this` indexed by a boolean ndarray.
+    void set_item_ndarray(
+        const nb::ndarray<bool, nb::c_contig>& key,
+        const std::variant<ARRAY_TYPE, scalar_variant_t<ARRAY_TYPE>>& val
+    )
+    {
+        if (key.ndim() == 0) {
+            std::string err_msg = fmt::format(
+                "{}.__setitem__: boolean key ndim = 0", ARRAY_TYPE::ARRAY_NAME
+            );
+            throw nb::index_error(err_msg.c_str());
+        } else if (key.ndim() > _ndim) {
+            std::string err_msg = fmt::format(
+                "{}.__setitem__: boolean key has too many dimensions for indexing; "
+                "self.ndim: {}, key.ndim: {} ",
+                ARRAY_TYPE::ARRAY_NAME,
+                _ndim,
+                key.ndim()
+            );
+            throw nb::index_error(err_msg.c_str());
+        }
+
+        for (std::size_t i = 0; i < key.ndim(); i++) {
+            if (_shape[i] != key.shape(i)) {
+                std::string err_msg = fmt::format(
+                    "{}.__setitem__: boolean key did not match shape along axis {}; "
+                    "self.shape[{}]: {}, key.shape[{}]: {}",
+                    ARRAY_TYPE::ARRAY_NAME,
+                    i,
+                    i,
+                    _shape[i],
+                    i,
+                    key.shape(i)
+                );
+                throw nb::index_error(err_msg.c_str());
+            }
+        }
+
+        if (std::holds_alternative<scalar_variant_t<ARRAY_TYPE>>(val)) {
+            auto&& scalar = std::get<scalar_variant_t<ARRAY_TYPE>>(val);
+            set_item_ndarray_from_scalar(key, scalar);
+        } else { /* std::holds_alternative<ARRAY_TYPE>(val) */
+            auto&& array = std::get<ARRAY_TYPE>(val);
+            set_item_ndarray_from_array(key, array);
+        }
+    }
+
     //! Python exported `__setitem__` method for APyArrays
     void set_item(
         const PyArrayKey_t& key,
         const std::variant<ARRAY_TYPE, scalar_variant_t<ARRAY_TYPE>>& val
     )
     {
+        if (std::holds_alternative<nb::ndarray<bool, nb::c_contig>>(key)) {
+            auto&& bool_key = std::get<nb::ndarray<bool, nb::c_contig>>(key);
+            set_item_ndarray(bool_key, val);
+            return; // early exit
+        }
+
         // Convert the variant of Python objects into a tuple of keys
         nb::tuple python_tuple_key;
         if (std::holds_alternative<nb::int_>(key)) {